exp queue: separate queue implementations into their own modules
